refactor: simplify package, add events from response to context
