+++ types/fixture/mock_module.go
@@ -1,4 +1,4 @@
-package server
+package fixture
 
 import (
 	"encoding/json"
