+++ types/fixture/module_key.go
@@ -1,17 +1,8 @@
-package server
+package fixture
 
 import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
 	"github.com/regen-network/regen-ledger/types"
 )
-
-type ModuleKey interface {
-	types.InvokerConn
-
-	ModuleID() types.ModuleID
-	Address() sdk.AccAddress
-}
 
 type InvokerFactory func(callInfo CallInfo) (types.Invoker, error)
