--- types/module/client/grpc_gateway/module.go
@@ -1,17 +1,0 @@
-package grpc_gateway
-
-/*
-import (
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	"github.com/regen-network/regen-ledger/types/module"
-)
-
-// Module is an interface that modules should implement to register grpc-gateway routes.
-type Module interface {
-	module.Module
-
-	RegisterGRPCGatewayRoutes(client.Context, *runtime.ServeMux)
-}
-
-*/
