+++ x/ecocredit/server/tests/bridge_test.go
@@ -6,6 +6,7 @@ import (
 	"strconv"
 	"testing"
 
+	sdkmodules "github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/gogo/protobuf/jsonpb"
 	"github.com/regen-network/gocuke"
 	"github.com/stretchr/testify/require"
@@ -14,9 +15,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/query"
 
 	"github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
-	"github.com/regen-network/regen-ledger/types"
-	"github.com/regen-network/regen-ledger/types/module"
-	"github.com/regen-network/regen-ledger/types/module/server"
+	"github.com/regen-network/regen-ledger/types/fixture"
 	"github.com/regen-network/regen-ledger/types/testutil"
 	"github.com/regen-network/regen-ledger/x/ecocredit"
 	"github.com/regen-network/regen-ledger/x/ecocredit/core"
@@ -43,14 +42,14 @@ func TestBridgeIntegration(t *testing.T) {
 func (s *bridgeSuite) Before(t gocuke.TestingT) {
 	s.t = t
 
-	ff := server.NewFixtureFactory(t, 2)
-	ff.SetModules([]module.Module{
+	ff := fixture.NewFixtureFactory(t, 2)
+	ff.SetModules([]sdkmodules.AppModule{
 		NewEcocreditModule(ff),
 	})
 
 	s.fixture = ff.Setup()
 	s.ctx = s.fixture.Context()
-	s.sdkCtx = s.ctx.(types.Context).WithContext(s.ctx)
+	s.sdkCtx = sdk.UnwrapSDKContext(s.ctx)
 
 	s.ecocreditServer = ecocreditServer{
 		MsgClient:   core.NewMsgClient(s.fixture.TxConn()),
@@ -72,7 +71,7 @@ func (s *bridgeSuite) BridgeServiceCallsBridgeReceiveWithMessage(a gocuke.DocStr
 
 	// reset context events
 	s.ctx = s.fixture.Context()
-	s.sdkCtx = s.ctx.(types.Context).WithContext(s.ctx)
+	s.sdkCtx = sdk.UnwrapSDKContext(s.ctx)
 
 	_, s.err = s.ecocreditServer.BridgeReceive(s.ctx, &msg)
 }
@@ -84,7 +83,7 @@ func (s *bridgeSuite) RecipientCallsBridgeWithMessage(a gocuke.DocString) {
 
 	// reset context events
 	s.ctx = s.fixture.Context()
-	s.sdkCtx = s.ctx.(types.Context).WithContext(s.ctx)
+	s.sdkCtx = sdk.UnwrapSDKContext(s.ctx)
 
 	_, s.err = s.ecocreditServer.Bridge(s.ctx, &msg)
 }
